+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## Unreleased
+### Added
+- `flipbox\element\lists\fields\ElementSourceList::$ignoreSearchKeywords` param to ignore recursively trying to generate keywords 
+
 ## 1.0.1 - 2018-09-20
 ### Added
 - `flipbox\element\lists\db\SourceUserElementQuery` to build a full query for users